+++ CHANGES.md
@@ -1,5 +1,6 @@
 # UNRELEASED
 - [FIXED] Case where `username` and `password` options were not used if a `url` was supplied.
+- [FIXED] Case where vcapServices was supplied with a basic-auth url
 
 # 2.3.0 (2018-06-08)
 - [FIXED] Removed addition of `statusCode` to response objects returned by promises.